refactor: Cambio de nombres en los controladores de Category y Testimony
